+++ Common/GameEngine/core/Time.cpp
@@ -16,16 +16,16 @@ namespace ecs {namespace core {
 
 	long Time::got(TimeUnit unit) {
 		if (unit == TimeUnit::Hours)
-			return (std::chrono::duration_cast<std::chrono::hours>(_clock.now().time_since_epoch()).count());
+			return (long) (std::chrono::duration_cast<std::chrono::hours>(_clock.now().time_since_epoch()).count());
 		else if (unit == TimeUnit::Minutes)
-			return (std::chrono::duration_cast<std::chrono::minutes>(_clock.now().time_since_epoch()).count());
+			return (long) (std::chrono::duration_cast<std::chrono::minutes>(_clock.now().time_since_epoch()).count());
 		else if (unit == TimeUnit::Seconds)
-			return (std::chrono::duration_cast<std::chrono::seconds>(_clock.now().time_since_epoch()).count());
+			return (long) (std::chrono::duration_cast<std::chrono::seconds>(_clock.now().time_since_epoch()).count());
 		else if (unit == TimeUnit::MilliSeconds)
-			return (std::chrono::duration_cast<std::chrono::milliseconds>(_clock.now().time_since_epoch()).count());
+			return (long) (std::chrono::duration_cast<std::chrono::milliseconds>(_clock.now().time_since_epoch()).count());
 		else if (unit == TimeUnit::MicroSeconds)
-			return (std::chrono::duration_cast<std::chrono::microseconds>(_clock.now().time_since_epoch()).count());
+			return (long) (std::chrono::duration_cast<std::chrono::microseconds>(_clock.now().time_since_epoch()).count());
 		else
-			return (std::chrono::duration_cast<std::chrono::nanoseconds>(_clock.now().time_since_epoch()).count());
+			return (long) (std::chrono::duration_cast<std::chrono::nanoseconds>(_clock.now().time_since_epoch()).count());
 	}
 }}
